Request 时间字段错误修改 RequestBody open 提供默认超时时间

+++ mqtt/src/main/kotlin/org/sheedon/mqtt/Request.kt
@@ -42,11 +42,11 @@ class Request internal constructor(
             super.backTopic(backTopic)
         }
 
-        override fun delayMilliSecond(delayMilliSecond: Int) = apply {
+        override fun delayMilliSecond(delayMilliSecond: Long) = apply {
             super.delayMilliSecond(delayMilliSecond)
         }
 
-        override fun delaySecond(delaySecond: Long) = apply {
+        override fun delaySecond(delaySecond: Int) = apply {
             super.delaySecond(delaySecond)
         }
